Fixed memory leak issue in normal use case. Added a logger which uses the -V flag in the run script, and added some parameters to the run script. Updated the readme to reflect the changes as well.

+++ software/run_unit_tests.sh
@@ -1,15 +1,35 @@
 #!/bin/sh
 # Script to setup git repos and other things needed for this repo. 
 
-# Setup software. 
-echo "Removing old executable"
-rm unit_tests/test_suite.exe
+# Parse args
+help=false
+clean=true
+verbose=false
 
-echo "Compiling tests"
-gcc -W unit_tests/test_suite.c unit_tests/unit_tests.c Unity/src/unity.c comSciCalc_lib/comscicalc.c comSciCalc_lib/comscicalc_operators.c -o unit_tests/test_suite
+usage() { echo "Usage: $0 [-V verbose] [-d dont clean]" 1>&2; exit 1; }
+
+while getopts ":Vd" flag
+do
+    case "${flag}" in
+        V) verbose=true;;
+        d) clean=false;;
+        *) usage;;
+    esac
+done
+
+if [ $clean == true ]; then
+    # Setup software. 
+    echo "Removing old executable"
+    rm unit_tests/test_suite.exe
+
+    echo "Compiling tests"
+    gcc -W -DVERBOSE=$verbose unit_tests/test_suite.c unit_tests/unit_tests.c Unity/src/unity.c comSciCalc_lib/comscicalc.c comSciCalc_lib/comscicalc_operators.c -o unit_tests/test_suite
+fi
 
 FILE=unit_tests/test_suite.exe
 if [ -f "$FILE" ]; then
     echo "Running tests"
     ./unit_tests/test_suite.exe
+else 
+    echo "No software to run. Try script without -d flag"
 fi
